+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow configuring gRPC keepalive
 - output identifier add in metric response object in `compute_plan perf` view.
 - Prevent use of `__` in asset metadata keys
+- Task input asset
 
 ### Fixed